anticipatory/requirements.txt
# Hotel Booking Evaluation System Dependencies

# Google Gemini API
google-genai>=1.0.0

# OpenAI API (optional - for OpenAI provider)
openai>=1.0.0

# HTTP/WebSocket clients
aiohttp>=3.9.0
websockets>=12.0

# Audio processing
numpy>=1.26.0

# Excel export with charts
openpyxl>=3.1.0

# LiveKit (for audio mode)
livekit>=0.12.0
livekit-api>=0.6.0
